* This commit fixes some warnings because of misprogramming. This commit   have some dirty hack (especially in fep/*). If some problem happens,   please revert them.  - gtk/test/entry.c: Add prototypes.  - gtk/caret-state-indicator.c:    + change #include order.    + (caret_state_indicator_update): Set value got from g_list_remove.  - fep/str.c    + strwidth (str_byte, nr_wchars): Change type to size_t from int.    + byte2width:      o (n, str_byte, nr_wchars): Change type to size_t from int.      o Fix bugs not allocating '\0' termination.      o Add dirty hack in mbsrtowcs caller because of glibc's stupid        implementation.    + byte2width:      o (n, str_byte, nr_wchars): Change type to size_t from int.      o Fix bugs not allocating '\0' termination.      o Add dirty hack in mbsrtowcs caller because of glibc's stupid        implementation.  - str.h: Sync with the str.c's change.  - fep/callbacks.c    + (strlen): Cast to int for comparing with signed int parameters.  - fep/escseq.c:    + (strlen): Ditto.  - helper/toolbar-common-gtk.c:    + (register_icon): Set value got from g_list_append.  - uim/uim-custom.c    + Remove redifinition of uim_custom_key_list_free. This prototype is already      defined in uim-custom.h).  - uim/uim-util.c:    + string_prefixp_internal: Add dirty hack for glibc's stupid implementation.    + iterate_lists:      - (rest, rests): Initialize as null list.  - uim/m17nlib.c: Surround #if 0 in unused functions.  - uim/plugin.c: Change substition of dlopen.  - uim/slib.c: Add dirty hack for glibc's stupid implementation.

+++ fep/callbacks.c
@@ -785,7 +785,7 @@ static void make_page_strs(void)
         assert(page_width == 0);
         next = TRUE;
                                                   /* 全角1文字の幅 */
-        if (s_max_width >= cand_label_width + strlen(":") + 2 + strlen(" ") + index_width) {
+        if (s_max_width >= cand_label_width + (int)strlen(":") + 2 + (int)strlen(" ") + index_width) {
           /* 候補 + インデックス */
 
           cand_width = s_max_width - index_width - strlen(" ");
@@ -804,7 +804,7 @@ static void make_page_strs(void)
           }
           cand_width -= strlen(" ");
           cand_width = strhead(cand_str, cand_width);
-          if (cand_width <= cand_label_width + strlen(":")) {
+          if (cand_width <= cand_label_width + (int)strlen(":")) {
             cand_width = 1;
             strcpy(cand_str, " ");
             s_candidate.cand_col[index] = UNDEFINED;
@@ -977,7 +977,7 @@ static void set_candidate(void)
   }
   s_candidate_str = tab2space(uim_candidate_get_cand_str(cand));
   cand_width = strwidth(s_candidate_str);
-  if (s_candidate_col + cand_width + strlen(" ") + index_width > s_max_width) {
+  if (s_candidate_col + cand_width + (int)strlen(" ") + index_width > s_max_width) {
     strhead(s_candidate_str, s_max_width - s_candidate_col - strlen(" ") - index_width);
   }
   uim_candidate_free(cand);
